Shell scripts, test

+++ generateTest.sh
@@ -1,8 +1,13 @@
+if [ -z "$1" ]
+  then
+    echo "No argument supplied"
+    return 1;
+fi
 rm -rf data
 mkdir -p data
 cd data || return 1;
 
-for i in 1 2 3 4 5 6 7 8
+for ((i = 1 ; i <= $1 ; i++))
 do
   python ../generator.py -f $i.txt -c $((i * 1000))
 done
